Confidence intervals for model parameters

Computes confidence intervals for parameters in a fitted gllvm model.

# S3 method for gllvm
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object
an object of class 'gllvm'.

parm
a specification of which parameters are to be given confidence intervals, a vector of names. If missing, all parameters are considered.

level
the confidence level. Scalar between 0 and 1.

...
not used.

Examples

## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
X <- as.matrix(antTraits$env[,1:2])
# Fit gllvm model
fit <- gllvm(y = y, X = X, family = poisson())
# 95 % confidence intervals for coefficients of X variables
confint(fit, level = 0.95, parm = "Xcoef")

#>                 cilow          ciup
#> Xcoef1   -0.063975620   0.091749142
#> Xcoef2   -0.745200795   0.421958083
#> Xcoef3    0.174387288   0.441004782
#> Xcoef4    0.044072277   0.161201072
#> Xcoef5    0.015237564   0.087898157
#> Xcoef6   -0.074688396   0.122495727
#> Xcoef7    0.092808583   0.332820238
#> Xcoef8   -0.190536365   0.434677455
#> Xcoef9   -0.022224953   0.905182858
#> Xcoef10  -0.043422570   0.044184697
#> Xcoef11   0.004495302   0.074132243
#> Xcoef12   0.007425512   0.165258718
#> Xcoef13  -0.080707960  -0.005414363
#> Xcoef14   0.047365181   0.645105468
#> Xcoef15  -0.011843572   0.021336338
#> Xcoef16  -0.037505464   0.069256123
#> Xcoef17  -0.047129137   0.060731723
#> Xcoef18   0.016396707   0.202514315
#> Xcoef19  -0.074692977   0.026553538
#> Xcoef20  -0.159206082   0.069991593
#> Xcoef21  -0.076556593   0.122417233
#> Xcoef22  -0.011013361   0.085402162
#> Xcoef23  -0.110225405   0.008091575
#> Xcoef24  -0.109713561   0.088524495
#> Xcoef25  -0.055386514   0.095062749
#> Xcoef26  -0.230324956   0.160468400
#> Xcoef27  -0.064303861   0.048439856
#> Xcoef28  -0.039570180   0.074979844
#> Xcoef29  -0.070052740   0.119560981
#> Xcoef30  -0.733236187   0.497600613
#> Xcoef31  -0.015371729   0.042069079
#> Xcoef32  -0.099992679   0.199477569
#> Xcoef33  -0.051456178   0.020762037
#> Xcoef34  -0.171197764   0.127488075
#> Xcoef35  -0.083277603   0.192588155
#> Xcoef36  -0.046642802  -0.004857005
#> Xcoef37  -0.112479311  -0.003737233
#> Xcoef38  -0.227850829   0.127575563
#> Xcoef39  -0.046680226   0.132710129
#> Xcoef40  -0.080419154   0.024332971
#> Xcoef41  -0.098976291   0.016781567
#> Xcoef42  -0.120062570   0.114515672
#> Xcoef43   0.247808409   2.122110503
#> Xcoef44   0.333573030   0.612426662
#> Xcoef45   0.069716223   0.207603028
#> Xcoef46  -0.189174429  -0.028186288
#> Xcoef47   0.067979617   0.272695476
#> Xcoef48   0.052609420   0.330441707
#> Xcoef49 -11.001460977 -10.999799388
#> Xcoef50   0.028130044   1.445650044
#> Xcoef51  -0.145409787   0.003318737
#> Xcoef52  -0.472162974  -0.135746609
#> Xcoef53  -0.045766700   0.180497744
#> Xcoef54  -0.054675261   0.046279979
#> Xcoef55   0.128959457   0.943680921
#> Xcoef56  -0.273528294  -0.140167006
#> Xcoef57  -0.338786325  -0.025014000
#> Xcoef58  -0.662142384   0.113681125
#> Xcoef59  -0.147715108   0.161283605
#> Xcoef60  -0.182201096   0.007487707
#> Xcoef61  -0.191305335   0.135823513
#> Xcoef62  -0.449465643   0.044662132
#> Xcoef63   0.057443076   0.180649309
#> Xcoef64  -2.060133398   0.171817000
#> Xcoef65  -0.322501221   0.062404268
#> Xcoef66 -32.380251547 -32.380202784
#> Xcoef67  -0.011189800   0.458266100
#> Xcoef68  -0.117415887   0.051279683
#> Xcoef69  -0.387588733  -0.050903260
#> Xcoef70 -29.360416941 -29.360253372
#> Xcoef71  -0.074751480   1.065182618
#> Xcoef72  -0.075586577   0.011994071
#> Xcoef73   0.041402189   0.348094612
#> Xcoef74  -0.060648928   0.040734873
#> Xcoef75  -0.173229128   0.244332662
#> Xcoef76  -0.069840613   0.257647134
#> Xcoef77  -0.128779391  -0.044788247
#> Xcoef78  -0.254573465  -0.019081387
#> Xcoef79  -0.102661689   0.238538978
#> Xcoef80   0.017092448   0.227093397
#> Xcoef81  -0.092635188   0.054344299
#> Xcoef82  -0.404329802  -0.018451154