Wood-decaying fungi data

Dataset of 1666 binary observations for 215 fungal species, in different 53 European Beech forests spread across 8 regions.

data(fungi)

Format

Y

A data frame with the presence-absences of 215 fungal species measured at 1666 logs.

X

A data frame of 8 predictor variables.

TR

A data frame of the traits used in Abrego et al. (2022).

tree

The phylogenetic tree.

C

The phylogenetic covariance matrix.

dist

The phylogenetic distance matrix.

Details

Observations of fungi species inhabiting European beech logs, in different European countries. The countries have been grouped in eight different regions. Logs were surveyed in 53 different reserves (or sites). Included environment and trait covariates are limited to those analyzed in the original article, though more are available in the published dataset on datadryad.org.

References

Abrego, N., Bässler, C., Christensen, M., and Heilmann‐Clausen, J. (2022). Traits and phylogenies modulate the environmental responses of wood‐inhabiting fungal communities across spatial scales. Journal of Ecology, 110(4), 784-798.

Abrego, N., Bässler, C., Christensen, M., and Heilmann-Clausen, J. (2022). Data and code from: Traits and phylogenies modulate the environmental responses of wood-inhabiting fungal communities across spatial scales [Dataset]. Dryad. https://doi.org/10.5061/dryad.t76hdr82r

Examples

if (FALSE) { # \dontrun{
data(fungi)
Y <- fungi$Y
X <- fungi$X
TR <- fungi$TR
C <- fungi$C
dist <- fungi$dist

#model <- gllvm(y = Y, X = cbind(int = 1, X), TR = TR, 
#        formula = ~DBH.CM + AVERDP + I(AVERDP^2) + CONNECT10 + TEMPR + PRECIP + 
#        log.AREA + (DBH.CM + AVERDP + I(AVERDP^2) + CONNECT10 + TEMPR + PRECIP + 
#        log.AREA):(FB.type + Sp.log.vol.µ3 + Lifestyle), 
#        family = "binomial", num.lv = 0, studyDesign = X[,c("REGION", "RESERVE")], 
#                     colMat = list(C, dist = dist), colMat.rho.struct = "term", 
#                     row.eff = ~(1 | REGION/RESERVE), sd.errors = FALSE, 
#                     randomX = ~int + DBH.CM + AVERDP + I(AVERDP^2) + 
#                                CONNECT10 + TEMPR + PRECIP + log.AREA, 
#                     beta0com = TRUE, nn.colMat = 10, maxit = 20000)
} # }