Plot Diagnostics for an gllvm Object

Five plots (selectable by which) are currently available: a plot of residuals against linear predictors of fitted values, a Normal Q-Q plot of residuals with a simulated point-wise 95% confidence interval envelope, residuals against row index and column index and scale location plot.

Usage

# S3 method for class 'gllvm'
plot(
  x,
  which = 1:5,
  caption = c("Residuals vs linear predictors", "Normal Q-Q", "Residuals vs row",
    "Residuals vs column", "Scale-Location"),
  var.colors = NULL,
  add.smooth = TRUE,
  envelopes = TRUE,
  reps = 150,
  envelope.col = c("blue", "lightblue"),
  n.plot = NULL,
  ...
)

Arguments

x

an object of class 'gllvm'.

which

if a subset of the plots is required, specify a subset of the numbers 1:5, see caption below.

caption

captions to appear above the plots.

var.colors

colors for responses, vector with length of number of response variables or 1. Defaults to NULL, when different responses have different colors.

add.smooth

logical indicating if a smoother should be added.

envelopes

logical, indicating if simulated point-wise confidence interval envelope will be added to Q-Q plot, defaults to TRUE

reps

number of replications when simulating confidence envelopes for normal Q-Q plot

envelope.col

colors for envelopes, vector with length of two

n.plot

number of species (response variables) to be plotted. Defaults to NULL when all response variables are plotted. Might be useful when data is very high dimensional.

...

additional graphical arguments.

Details

plot.gllvm is used for model diagnostics. Dunn-Smyth residuals (randomized quantile residuals) (Dunn and Smyth, 1996) are used in plots. Colors indicate different species.

References

Dunn, P. K., and Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236-244.

Hui, F. K. C., Taskinen, S., Pledger, S., Foster, S. D., and Warton, D. I. (2015). Model-based approaches to unconstrained ordination. Methods in Ecology and Evolution, 6:399-411.

See also

Author

Jenni Niku <jenni.m.e.niku@jyu.fi>

Examples