Plot random slope coefficients

Plots random slopes and their prediction intervals.

Usage

# S3 method for class 'gllvm'
randomCoefplot(
  object,
  y.label = TRUE,
  which.Xcoef = NULL,
  cex.ylab = 0.5,
  mfrow = NULL,
  mar = c(4, 6, 2, 1),
  xlim.list = NULL,
  order = FALSE,
  ...
)

Arguments

object

an object of class 'gllvm'.

y.label

logical, if TRUE (default) colnames of y with respect to coefficients are added to plot.

which.Xcoef

fector indicating which covariate coefficients will be plotted. Can be vector of covariate names or numbers. Default is NULL when all covariate coefficients are plotted.

cex.ylab

the magnification to be used for axis annotation relative to the current setting of cex.

mfrow

same as mfrow in par. If NULL (default) it is determined automatically.

mar

vector of length 4, which defines the margin sizes: c(bottom, left, top, right). Defaults to c(4,5,2,1).

xlim.list

list of vectors with length of two to define the intervals for x axis in each covariate plot. Defaults to NULL when the interval is defined by the range of point estimates and confidence intervals

order

logical, if TRUE (default), coefficients are sorted according to the point estimates

...

additional graphical arguments.

Author

Jenni Niku <jenni.m.e.niku@jyu.fi>, Francis K.C. Hui, Bert van der Veen, Sara Taskinen,

Examples

if (FALSE) { # \dontrun{
## Load a dataset from the mvabund package
data(antTraits, package = "mvabund")
y <- as.matrix(antTraits$abund)
X <- as.matrix(antTraits$env)
TR <- antTraits$traits
# Fit model with random slopes
fitF <- gllvm(y = y, X = X, TR = TR,
 formula = ~ Bare.ground + Bare.ground : Webers.length,
 family = poisson(), randomX = ~ Bare.ground)
randomCoefplot(fitF)
} # }