Dunn-Smyth residuals for gllvm model

Calculates Dunn-Smyth residuals for gllvm model.

# S3 method for gllvm
residuals(object, ...)

Arguments

object

an object of class 'gllvm'.

...

not used.

Value

residuals

matrix of residuals

linpred

matrix of linear predictors

Details

Computes Dunn-Smyth residuals (randomized quantile residuals, Dunn and Smyth, 1996) for gllvm model. For the observation \(Y_{ij}\) Dunn-Smyth residuals are defined as

$$r_{ij}=\Phi^{-1}(u_{ij}F_{ij}(y_{ij}) + (1-u_{ij})F_{ij}^-(y_{ij})),$$

where \(\Phi(.)\) and \(F_{ij}(.)\) are the cumulative probability functions of the standard normal distribution, \(F_{ij}^-(y))\) is the limit as \(F_{ij}(y)\) is approached from the negative side, and \(u_{ij}\) has been generated at random from the standard uniform distribution.

References

Dunn, P. K., and Smyth, G. K. (1996). Randomized quantile residuals. Journal of Computational and Graphical Statistics, 5, 236-244.

Hui, F. K. C., Taskinen, S., Pledger, S., Foster, S. D., and Warton, D. I. (2015). Model-based approaches to unconstrained ordination. Methods in Ecology and Evolution, 6:399-411.

Author

Jenni Niku <jenni.m.e.niku@jyu.fi>

Examples

if (FALSE) {
# Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = poisson())
# residuals
res <- residuals(fit)
}