Summarizing gllvm model fits

A summary of the fitted 'gllvm' object, including function call, distribution family and model parameters.

# S3 method for gllvm
summary(
  object,
  digits = max(3L, getOption("digits") - 3L),
  signif.stars = getOption("show.signif.stars"),
  dispersion = FALSE,
  spp.intercepts = FALSE,
  row.intercepts = FALSE,
  Lvcoefs = FALSE,
  rotate = FALSE,
  type = NULL,
  ...
)

# S3 method for summary.gllvm
print(x, ...)

Arguments

object

an object of class 'gllvm'

digits

the number of significant digits to use when printing

signif.stars

If TRUE, significance stars are printed for each coefficient, defaults to TRUE

dispersion

option to return dispersion parameters, defaults to FALSE

spp.intercepts

option to return species intercepts, defaults to FALSE

row.intercepts

option to return row intercepts, defaults to FALSE

Lvcoefs

option to return species scores in the ordination, defaults to FALSE. Returns species optima for quadratic model.

rotate

defaults to FALSE. If TRUE rotates the output of the latent variables to principal direction, so that it coincides with the ordiplot results. If both unconstrained and constrained latent variables are included, predictor slopes are not rotated.

type

to match "type" in ordiplot.gllvm

...

not used.

x

a summary object

Details

Various options are available to include extra parameter estimates in the summary, which have been excluded by default, for readability.

Author

Jenni Niku <jenni.m.e.niku@jyu.fi>, Bert van der Veen

Examples

if (FALSE) {
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = poisson())
summary(fit)
}