Confidence intervals for model parameters

Computes confidence intervals for parameters in a fitted gllvm model.

# S3 method for gllvm
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object
an object of class 'gllvm'.

parm
a specification of which parameters are to be given confidence intervals, a vector of names. If missing, all parameters are considered.

level
the confidence level. Scalar between 0 and 1.

...
not used.

Examples

## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
X <- as.matrix(antTraits$env[,1:2])
# Fit gllvm model
fit <- gllvm(y = y, X = X, family = poisson())
# 95 % confidence intervals for coefficients of X variables
confint(fit, level = 0.95, parm = "Xcoef")

#>                 cilow          ciup
#> Xcoef1   -0.063821480   0.105996181
#> Xcoef2   -0.569303324   0.475607368
#> Xcoef3   -0.016637745   0.473207083
#> Xcoef4    0.041736286   0.115784929
#> Xcoef5    0.013925743   0.080693898
#> Xcoef6   -0.046927276   0.112013688
#> Xcoef7    0.087306565   0.212025624
#> Xcoef8   -0.245457970   0.368350452
#> Xcoef9   -0.097730680   0.611761048
#> Xcoef10  -0.046865448   0.051529060
#> Xcoef11   0.001421645   0.069175063
#> Xcoef12   0.005155421   0.151333374
#> Xcoef13  -0.081742318  -0.010551501
#> Xcoef14  -0.001763957   0.171408656
#> Xcoef15  -0.012762632   0.021449797
#> Xcoef16  -0.037600716   0.053977978
#> Xcoef17  -0.048950996   0.060843312
#> Xcoef18   0.013060723   0.191527135
#> Xcoef19  -0.073662324   0.020451146
#> Xcoef20  -0.153308682   0.045042021
#> Xcoef21  -0.081052017   0.109757628
#> Xcoef22  -0.015139382   0.073261854
#> Xcoef23  -0.115063574   0.005522525
#> Xcoef24  -0.113906133   0.072547729
#> Xcoef25  -0.054213590   0.095728879
#> Xcoef26  -0.163586865   0.061242369
#> Xcoef27  -0.071084007   0.036832165
#> Xcoef28  -0.045774270   0.063026603
#> Xcoef29  -0.073107003   0.106172006
#> Xcoef30  -0.076376920   0.187296073
#> Xcoef31  -0.016932789   0.037423396
#> Xcoef32  -0.043487686   0.152364349
#> Xcoef33  -0.051886342   0.025111960
#> Xcoef34  -0.222970236   0.087283270
#> Xcoef35  -0.080996219   0.191990946
#> Xcoef36  -0.047159957  -0.004979604
#> Xcoef37  -0.107236037  -0.004613889
#> Xcoef38  -0.235091383   0.111972889
#> Xcoef39  -0.055291263   0.163744712
#> Xcoef40  -0.081322081   0.037246975
#> Xcoef41  -0.098673938   0.011895837
#> Xcoef42  -0.119137842   0.126597236
#> Xcoef43   0.120145443   1.216068826
#> Xcoef44  -0.007065100   0.670223002
#> Xcoef45   0.076959934   0.163544385
#> Xcoef46  -0.184115219  -0.033358726
#> Xcoef47   0.071624398   0.252519829
#> Xcoef48   0.059792814   0.226548775
#> Xcoef49  -3.975879518  -3.970177658
#> Xcoef50  -0.106477302   0.993879450
#> Xcoef51  -0.129264777   0.022859667
#> Xcoef52  -0.353629369  -0.084776839
#> Xcoef53  -0.048421263   0.169601342
#> Xcoef54  -0.055331391   0.040225287
#> Xcoef55   0.028511021   0.272212965
#> Xcoef56  -0.248281779  -0.123750206
#> Xcoef57  -0.274237458  -0.017993173
#> Xcoef58  -0.704558134   0.118528744
#> Xcoef59  -0.145770032   0.144715342
#> Xcoef60  -0.173634372   0.006102906
#> Xcoef61  -0.178584709   0.105287108
#> Xcoef62  -0.366393407   0.040228602
#> Xcoef63   0.040652067   0.154122736
#> Xcoef64  -1.964852373   0.458130818
#> Xcoef65  -0.287261977   0.055375145
#> Xcoef66 -29.664122051 -29.664001851
#> Xcoef67  -0.015825182   0.210545335
#> Xcoef68  -0.120644249   0.044007779
#> Xcoef69  -0.417342012  -0.047278272
#> Xcoef70 -27.815264773 -27.814882460
#> Xcoef71   0.078342770   0.384834054
#> Xcoef72  -0.076890174   0.007367301
#> Xcoef73   0.067101276   0.291804648
#> Xcoef74  -0.063537524   0.043826163
#> Xcoef75  -0.222024075   0.165299055
#> Xcoef76  -0.072356809   0.234537045
#> Xcoef77  -0.130517517  -0.045144445
#> Xcoef78  -0.224814286  -0.013104062
#> Xcoef79  -0.109457921   0.223498042
#> Xcoef80  -0.011571350   0.249053821
#> Xcoef81  -0.091227898   0.069541635
#> Xcoef82  -0.358661265  -0.011441966