Confidence intervals for model parameters

Computes confidence intervals for parameters in a fitted gllvm model.

# S3 method for gllvm
confint(object, parm = NULL, level = 0.95, ...)

Arguments

object
an object of class 'gllvm'.

parm
a specification of which parameters are to be given confidence intervals, a vector of names. Examples of options are "beta0", "Xcoef",theta", "phi". If missing, all parameters are considered.

level
the confidence level. Scalar between 0 and 1.

...
not used.

Author

Jenni Niku <jenni.m.e.niku@jyu.fi>

Examples

if (FALSE) {
## Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
X <- as.matrix(antTraits$env[,1:2])
# Fit gllvm model
fit <- gllvm(y = y, X = X, family = poisson())
# 95 % confidence intervals for coefficients of X variables
confint(fit, level = 0.95, parm = "Xcoef")
}