Extract residual correlations from gllvm object

Calculates the residual correlation matrix for gllvm model.

# S3 method for gllvm
getResidualCor(object, adjust = 1)

Arguments

object
an object of class 'gllvm'.

adjust
The type of adjustment used for negative binomial and binomial distribution when computing residual correlation matrix. Options are 0 (no adjustment), 1 (the default adjustment) and 2 (alternative adjustment for NB distribution). See details.

Details

Residual correlation matrix is calculated based on the residual covariance matrix, see details from getResidualCov.gllvm.

Author

Francis K.C. Hui, Jenni Niku, David I. Warton

Examples

#'# Extract subset of the microbial data to be used as an example
data(microbialdata)
y <- microbialdata$Y[, order(colMeans(microbialdata$Y > 0), 
                     decreasing = TRUE)[21:40]]
fit <- gllvm(y, family = poisson())
fit$logL

#> [1] -4242.667

cr <- getResidualCor(fit)
cr[1:5,1:5]

#>             OTU_79    OTU_164    OTU_462     OTU_95    OTU_833
#> OTU_79   1.0000000  0.3893696  0.3428408 -0.7496213  0.9956260
#> OTU_164  0.3893696  1.0000000  0.9987499 -0.9015132  0.4737217
#> OTU_462  0.3428408  0.9987499  1.0000000 -0.8787543  0.4291073
#> OTU_95  -0.7496213 -0.9015132 -0.8787543  1.0000000 -0.8081796
#> OTU_833  0.9956260  0.4737217  0.4291073 -0.8081796  1.0000000

if (FALSE) {
# Load a dataset from the mvabund package
data(antTraits)
y <- as.matrix(antTraits$abund)
# Fit gllvm model
fit <- gllvm(y = y, family = poisson())
# residual correlations:
cr <- getResidualCor(fit)
# Plot residual correlations:
install.packages("corrplot", "gclus")
library(corrplot)
library(gclus)
corrplot(cr[order.single(cr), order.single(cr)], diag = F,
  type = "lower", method = "square", tl.cex = 0.8, tl.srt = 45, tl.col = "red")
  }