Simulate data from gllvm fit

Generate new data using the fitted values of the parameters

Usage

# S3 method for class 'gllvm'
simulate(object, nsim = 1, seed = NULL, conditional = FALSE, ...)

Arguments

object

an object of class 'gllvm'.

nsim

an optional positive integer specifying the number of simulated datasets. Defaults to 1.

seed

an optional integer to set seed number, passed to set.seed. Defaults to a random seed number.

conditional

if conditional = FALSE simulates marginally over the latent variables.

...

not used.

Value

A matrix containing generated data.

Details

simulate function for gllvm objects.

Author

David Warton, Jenni Niku <jenni.m.e.niku@jyu.fi>

Examples

 # \donttest{
# Load a dataset from the mvabund package
data(antTraits, package = "mvabund")
y <- as.matrix(antTraits$abund)
X <- scale(antTraits$env[, 1:3])
# Fit gllvm model
fit <- gllvm(y = y, X, family = poisson())
# Simulate data
newdata <- simulate(fit)
#> Error in terms.formula(object, data = data): argument is not a valid model
# }